chore(client): Add views, components and refactor providers
