+++ packages/client/src/components/Button.tsx
@@ -1,0 +1,1 @@
+export function Button() {}
