+++ packages/client/src/components/SelectInput.tsx
@@ -1,0 +1,1 @@
+export function SelectInput() {}
