+++ packages/client/src/components/TextInput.tsx
@@ -1,0 +1,1 @@
+export function TextInput() {}
