+++ packages/client/src/index.ts
@@ -1,1 +1,7 @@
-export { AppProvider } from './providers/app.provider';
+// Providers
+export { AppProvider } from './providers/AppProvider';
+
+// Views
+export { PasswordResetView } from './views/password_reset/PasswordReset';
+export { SignInView } from './views/signin/SignIn';
+export { SignUpView } from './views/signup/SignUp';
