+++ packages/client/src/views/password_reset/PasswordReset.tsx
@@ -1,0 +1,1 @@
+export function PasswordResetView() {}
