+++ packages/client/src/views/signin/SignIn.tsx
@@ -1,0 +1,1 @@
+export function SignInView() {}
