+++ packages/client/src/views/signup/SignUp.tsx
@@ -1,0 +1,1 @@
+export function SignUpView() {}
